Separate save offline bundle to its own class
